refactor: move and rename interface aggregation tck/int tests
