+++ packages/graphql/tests/tck/experimental/aggregations/aggregation-interfaces-top-level.test.ts
@@ -19,8 +19,8 @@
 
 import type { DocumentNode } from "graphql";
 import { gql } from "graphql-tag";
-import { Neo4jGraphQL } from "../../../src";
-import { formatCypher, formatParams, translateQuery } from "../utils/tck-test-utils";
+import { Neo4jGraphQL } from "../../../../src";
+import { formatCypher, formatParams, translateQuery } from "../../utils/tck-test-utils";
 
 describe("Top level aggregation interfaces", () => {
     let typeDefs: DocumentNode;
